ref(macro): split js (core) and jsx (react) macro
