+++ packages/macro-lib/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@lingui/macro-lib",
+  "version": "3.17.1",
+  "description": "Utils for lingui macro",
+  "main": "./build/index.js",
+  "types": "./index.d.ts",
+  "scripts": {
+    "test:tsd": "tsd"
+  },
+  "license": "MIT",
+  "keywords": [],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/lingui/js-lingui.git"
+  },
+  "bugs": {
+    "url": "https://github.com/lingui/js-lingui/issues"
+  },
+  "engines": {
+    "node": ">=14.0.0"
+  },
+  "files": [
+    "LICENSE",
+    "README.md",
+    "build/"
+  ],
+  "dependencies": {
+    "@babel/runtime": "^7.20.13"
+  },
+  "devDependencies": {
+    "@lingui/conf": "^3.17.1",
+    "@babel/traverse": "^7.20.13",
+    "@babel/types": "^7.20.7"
+  },
+  "tsd": {
+    "compilerOptions": {
+      "strict": false
+    }
+  }
+}
