+++ packages/macro-lib/src/icu.test.ts
@@ -1,9 +1,9 @@
-import ICUMessageFormat, { Token } from "./icu"
+import { ICUMessageFormat, Token } from "./icu"
 import { Identifier } from "@babel/types"
 
 describe("ICU MessageFormat", function () {
   it("should collect text message", function () {
-    const messageFormat = new ICUMessageFormat()
+    const messageFormat = new ICUMessageFormat({} as any)
     const tokens: Token[] = [
       {
         type: "text",
@@ -19,7 +19,7 @@ describe("ICU MessageFormat", function () {
   })
 
   it("should collect text message with arguments", function () {
-    const messageFormat = new ICUMessageFormat()
+    const messageFormat = new ICUMessageFormat({} as any)
     const tokens: Token[] = [
       {
         type: "text",
