+++ packages/macro-lib/src/types.ts
@@ -1,0 +1,7 @@
+import type { LinguiConfigNormalized } from "@lingui/conf"
+
+export type LinguiMacroOpts = {
+  // explicitly set by CLI when running extraction process
+  extract?: boolean
+  linguiConfig?: LinguiConfigNormalized
+}
