+++ packages/macro/index.js
@@ -1,1 +1,1 @@
-export { default } from "./src"
+export { default } from "@lingui/core/src/macro"
