--- packages/macro/src/utils.ts
@@ -1,4 +1,0 @@
-export const makeCounter =
-  (index = 0) =>
-  () =>
-    index++
